+++ src/command/mod.rs
@@ -139,6 +139,7 @@ pub mod status;
 pub(crate) mod status_untracked;
 pub(crate) mod status_untracked_paths;
 pub mod switch;
+pub mod upgrade;
 pub mod web_assets;
 pub mod write_tree;